pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.4",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.10",
    "langsmith>=0.7.9",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
]
